Create Activity
Added the neccessary logic to create and Activity from a DTO (service,
mappers, controller code with validations).
Created the neccessary DTOs to return and get as request body.
Modified the ModelMapper to let it map the nested entity (projectDto).

Status: Finished

Observations: Only tests are left, but will do when the controller is
finished.