Activity Read by Id Project
Added the endpoint to receive all the activities of certain project by
its id.
Added the neccessary method to the Service Interface, Service
Implementation and Repository.
Also changed the validation for predecessors in post method to allow
only nulls and Strings that match the pattern.

Status: Finished

Observations: Integration Tests are left for this endpoint. Will do when
finishing the Activities Controller.